refactor: 알림 기준 서비스 클래스 통합
- 알림 기준 서비스 클래스를 하나로 통합
- 코드 개선 : build 정적 메서드 활용 -> 객체 생성 복잡도 감소
