+++ src/main/java/com/alzzaipo/notification/application/port/dto/NotificationCriterionView.java
@@ -1,5 +1,6 @@
 package com.alzzaipo.notification.application.port.dto;
 
+import com.alzzaipo.notification.domain.criterion.NotificationCriterion;
 import lombok.Getter;
 
 @Getter
@@ -9,9 +10,15 @@ public class NotificationCriterionView {
     private final int competitionRate;
     private final int lockupRate;
 
-    public NotificationCriterionView(Long uid, int competitionRate, int lockupRate) {
-        this.uid = uid.toString();
+    private NotificationCriterionView(String uid, int competitionRate, int lockupRate) {
+        this.uid = uid;
         this.competitionRate = competitionRate;
         this.lockupRate = lockupRate;
+    }
+
+    public static NotificationCriterionView build(NotificationCriterion notificationCriterion) {
+        return new NotificationCriterionView(notificationCriterion.getNotificationCriterionUID().toString(),
+            notificationCriterion.getMinCompetitionRate(),
+            notificationCriterion.getMinLockupRate());
     }
 }
